+++ src/analytics.ts
@@ -1,57 +1,10 @@
 import { Redis } from "@upstash/redis";
-
-export type Event = {
-  time?: number;
-  [key: string]: string | number | boolean | undefined;
-};
-
-class Key {
-  constructor(public readonly prefix: string, public readonly table: string, public readonly bucket: number) {}
-
-  public toString() {
-    return [this.prefix, this.table, this.bucket].join(":");
-  }
-  static fromString(key: string) {
-    const [prefix, table, bucket] = key.split(":");
-    return new Key(prefix, table, parseInt(bucket));
-  }
-}
-
-export type Window = `${number}${"s" | "m" | "h" | "d"}`;
-
-export type AnalyticsConfig = {
-  redis: Redis;
-  /**
-   * Configure the bucket size for analytics. All events inside the window will be stored inside
-   * the same bucket. This reduces the number of keys that need to be scanned when aggregating
-   * and reduces your cost.
-   *
-   * Must be either a string in the format of `1s`, `2m`, `3h`, `4d` or a number of milliseconds.
-   */
-  window: Window | number;
-  prefix?: string;
-
-  /**
-   * Configure the retention period for analytics. All events older than the retention period will
-   * be deleted. This reduces the number of keys that need to be scanned when aggregating.
-   *
-   * Can either be a string in the format of `1s`, `2m`, `3h`, `4d` or a number of milliseconds.
-   * 0, negative or undefined means that the retention is disabled.
-   *
-   * @default Disabled
-   *
-   * Buckets are evicted when they are read, not when they are written. This is much cheaper since
-   * it only requires a single command to ingest data.
-   */
-  retention?: Window | number
-  
-  /**
-   * Whether to initialize a cache. Cache is used in the `loadBuckets` method, which in turn is
-   * called in `count`, `aggregateBy` and `query` methods. This means that to only use `ingest`,
-   * there is no need to use cache.
-   */
-  cache?: boolean
-};
+import {
+  Event,
+  Key,
+  Window,
+  AnalyticsConfig
+} from "./types"
 
 class Cache<TValue> {
   private readonly cache: Map<string, { value: TValue; createdAt: number }>;
